pyproject.toml
[tool.poetry]
name = "llm-zoomcamp"
version = "0.1.0"
description = "LLM Zoomcamp course from DataTalks.club"
authors = ["Vladislav Garist <vv.garist@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.4"
notebook = "^7.2.1"
openai = "^1.35.3"
elasticsearch = "^8.14.0"
scikit-learn = "^1.5.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"
tiktoken = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
